Fixed some syntactic bugs, added more verbose output

+++ hwtest/launch.sh
@@ -3,9 +3,6 @@
 LIST_PRODUCERS=listProducers.txt
 LIST_MERGERS=listMergers.txt
 
-MERGERID=lxplus0095
-CATID=lxplus0095
-
 LUMI_LENGTH_MEAN=20
 LUMI_LENGTH_SIGMA=0.01
 
@@ -25,12 +22,13 @@ source $TEST_BASE/hwtest/tools.sh
 
 #-------------------------------------------------------------------------------
 function launch_main {
+    echo "+ Launching the test ..."
     kill_previous_mergers
-    delete_previous_runs
+    # delete_previous_runs
 
-    launch_producers run100.cfg 1
-    sleep $LUMI_LENGTH_MEAN
-    launch_merger 100 optionC wbua-TME-ComputeNode7 1
+    # launch_producers run100.cfg 1
+    # sleep $LUMI_LENGTH_MEAN
+    # launch_merger 100 optionC wbua-TME-ComputeNode7 1
     #launch_simple_cat 300 0 lxplus0095
     
 #    launch_producers mergeConfigTest
@@ -40,23 +38,43 @@ function launch_main {
 #    launch_producers run400.cfg
 #    launch_producers run500.cfg
 
+    echo "+ ... Finished launching the test."
 } # launch_main
+
 
 #-------------------------------------------------------------------------------
 function kill_previous_mergers {
-
-    for NODE in `cat $LIST_MERGERS | cut -d' ' -f1`; do
-        COMMAND=$(cat <<'EOF'
+    echo "++ Killing previous mergers ..."
+    for NODE in $(parse_machine_list $LIST_MERGERS); do
+        COMMAND="$(cat <<'EOF'
             PS_LINE=$(ps awwx | grep python | egrep -v "grep|bash");\
-            PID=$(echo $PS_LINE | awk '{print $1}');\
-            kill -9 $PID
+            PID=$(echo $PS_LINE | awk '{print $1}')                ;\
+            if [[ ! -z "$PID" ]]; then                              \
+                kill -9 $PID                                       ;\
+            fi
 EOF
-        )
-        echo $NODE
-        echo "      $COMMAND"
-        ssh  $NODE "$COMMAND"
+            )"
+        echo_and_ssh $NODE "$COMMAND"
     done
+    echo "++ ... Finished killing previous mergers."
 } # kill_previous_mergers
+
+
+# #-------------------------------------------------------------------------------
+# function kill_previous_mergers {
+#     for NODE in $(parse_machine_list $LIST_MERGERS); do
+#         COMMAND=$(cat <<'EOF'
+#             PS_LINE=$(ps awwx | grep python | egrep -v "grep|bash");\
+#             PID=$(echo $PS_LINE | awk '{print $1}');\
+#             if [[ ! -z "$PID" ]]; then\
+#                 kill -9 $PID;\
+#             fi\
+# EOF
+#         )
+#         #echo "$COMMAND"
+#         echo_and_ssh $NODE "$COMMAND"
+#     done
+# } # kill_previous_mergers
 
 #-------------------------------------------------------------------------------
 # Expects the run number as the first argument, 
@@ -68,44 +86,46 @@ function launch_merger {
     export NODE=${3:-}
     if [ -z $NODE ]; then
         ## Iterate over all nodes in the list given in the $LIST_PRODUCERS
-        for NODE in $(parse_list_file $LIST_MERGERS); do
+        for NODE in $(parse_machine_list $LIST_MERGERS); do
             launch_merger $RUN $THEOPTION $NODE
         done
     else
-        echo $NODE
-        mkdir -p $TEST_BASE/hwtest/${NODE}
-        CONFIG=$TEST_BASE/hwtest/dataFlowMerger.conf
-        cp $TEST_BASE/dataFlowMergerTemplate.conf $CONFIG
-        sed -e "s|AAA|$INPUT_LOCATION/${NODE}/unmergedDATA/run${RUN}|" \
-            -e "s|BBB|$INPUT_LOCATION/${NODE}/unmergedMON|"            \
-            -e "s|OPTION|$THEOPTION|"                                  \
-            -e "s|CCC|$OUTPUT_LOCATION/mergerMini|"                    \
-            -e "s|DDD|$OUTPUT_LOCATION/mergerMacro|"                   \
-            -e "s|LOG|$ROOT_LOCATION/logFormat.conf|"                  \
-            -i $CONFIG
-
-        # sed -i "s|dataFlowMerger.conf|$TEST_BASE/hwtest/${NODE}/dataFlowMerger.conf|" $TEST_BASE/hwtest/${NODE}/dataFlowMergerInLine;
-        # sed -i "s|dataFlowMerger.conf|$TEST_BASE/hwtest/${NODE}/dataFlowMerger.conf|" $TEST_BASE/hwtest/${NODE}/Logging.py;
-
-        ## Make sure that the remote folder to contain source code exists
-        COMMAND="$(cat << EOF
-        if [[ ! -d $ROOT_LOCATION/$NODE ]]; then \
-            mkdir -p $ROOT_LOCATION/$NODE; \
-        fi
-EOF)"
+#         echo $NODE
+#         mkdir -p $TEST_BASE/hwtest/${NODE}
+#         CONFIG=$TEST_BASE/hwtest/dataFlowMerger.conf
+#         cp $TEST_BASE/dataFlowMergerTemplate.conf $CONFIG
+#         sed -e "s|AAA|$INPUT_LOCATION/${NODE}/unmergedDATA/run${RUN}|" \
+#             -e "s|BBB|$INPUT_LOCATION/${NODE}/unmergedMON|"            \
+#             -e "s|OPTION|$THEOPTION|"                                  \
+#             -e "s|CCC|$OUTPUT_LOCATION/mergerMini|"                    \
+#             -e "s|DDD|$OUTPUT_LOCATION/mergerMacro|"                   \
+#             -e "s|LOG|$ROOT_LOCATION/logFormat.conf|"                  \
+#             -i $CONFIG
+# 
+#         # sed -i "s|dataFlowMerger.conf|$TEST_BASE/hwtest/${NODE}/dataFlowMerger.conf|" $TEST_BASE/hwtest/${NODE}/dataFlowMergerInLine;
+#         # sed -i "s|dataFlowMerger.conf|$TEST_BASE/hwtest/${NODE}/dataFlowMerger.conf|" $TEST_BASE/hwtest/${NODE}/Logging.py;
+# 
+#         ## Make sure that the remote folder to contain source code exists
+#         echo_and_ssh $NODE "mkdir -p $ROOT_LOCATION/$NODE"
+# 
+#         ## Sync the source code to the remote node
+#         rsync -aW $TEST_BASE/ $NODE:$ROOT_LOCATION/$NODE
+# 
+#         ## Launch the merger
+#         COMMAND="$(cat << EOF
+#         (   cd $OUTPUT_LOCATION ; \
+#             nohup $ROOT_LOCATION/${NODE}/dataFlowMergerInLine \
+#         )   >& $ROOT_LOCATION/hwtest/merger_${THEOPTION}_run${RUN}_${NODE}.log &
+# EOF
+        COMMAND="$(cat << EOHD
+        (                                                        \
+            cd $OUTPUT_LOCATION ;                                \
+            nohup $ROOT_LOCATION/${NODE}/dataFlowMergerInLine    \
+        )
+EOHD
+        )"
         echo_and_ssh $NODE $COMMAND
-
-        ## Sync the source code to the remote node
-        rsync -aW $TEST_BASE/ $NODE:$ROOT_LOCATION/$NODE
-
-        ## Launch the merger
-        COMMAND="$(cat << EOF
-        (   cd $OUTPUT_LOCATION ; \
-            nohup $ROOT_LOCATION/${NODE}/dataFlowMergerInLine \
-        )   >& $ROOT_LOCATION/hwtest/merger_${THEOPTION}_run${RUN}_${NODE}.log &
-EOF)"
-        echo_and_ssh $NODE $COMMAND
-    fi # [ -z $NODE ]
+    fi
 } # launch_merger
 
 
@@ -116,7 +136,7 @@ function launch_producers {
     CONFIG=${1:-mergeConfigForReal}
     DOSUBFOLDER=${2:-0}
     TOTALBUS=$(count_args $LIST_PRODUCERS)
-    for NODE in $(parse_list_file $LIST_PRODUCERS); do
+    for NODE in $(parse_machine_list $LIST_PRODUCERS); do
         echo $NODE
         rsync -aW $TEST_BASE/ $NODE:$ROOT_LOCATION/;
         SUBFOLDER=""
@@ -145,7 +165,7 @@ function launch_simple_cat {
     ## The number of process per node is passed as the first arg, default=1
     RUN=${1:-300}
     LS=${2:-1}
-    NODE=${3:-$CATID}
+    NODE=$3
     STREAM=A
     echo $RUN
     SOURCE_BASE=$INPUT_LOCATION/unmergedDATA/run${RUN}
@@ -175,26 +195,31 @@ function delete_previous_runs {
 
 #-------------------------------------------------------------------------------
 function echo_and_ssh {
-    $NODE=$1
-    $COMMAND=$2
-    echo "    $COMMAND"
+    NODE=$1
+    COMMAND="$2"
+    echo "+++ $NODE"
+    ## Format the command for printing
+    FORMATTED_COMMAND="$(echo $COMMAND | tr ';' '\n'| sed -E 's/^/    /g')"
+    echo "$FORMATTED_COMMAND"
     ssh $NODE "$COMMAND"
 }  ## echo_and_ssh
 
 
 #-------------------------------------------------------------------------------
-parse_list_file {
+function parse_machine_list {
     ## sed removes Bash/Python-style comments starting with `#'
     ## awk makes sure to ignore white space around the node name
-    echo "$(sed 's/#.*$//' | awk '{print $1}' $1)"
-} ## parse_list_file
+    echo "$(sed 's/#.*$//' $1 | awk '{print $1}')"
+} ## parse_machine_list
 
 
 #-------------------------------------------------------------------------------
 function echo_and_rm {
-    echo "Deleting $@ ..."
-    rm -rf $@
-    echo "... done."
+    if [[ ! -z "$@" ]]; then
+        echo "Deleting $@ ..."
+        rm -rf $@
+        echo "... done."
+    fi
 } # echo_and_rm
 
 launch_main
